requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.1.3
django-admin-interface==0.26.0
django-colorfield==0.12.0
django-cors-headers==4.7.0
django-filter==25.1
django-import-export==4.3.7
django-tinymce==4.1.0
djangorestframework==3.15.2
pillow==11.0.0
python-slugify==8.0.4
sqlparse==0.5.2
tablib==3.8.0
text-unidecode==1.3
tzdata==2025.1
